Async authentication layer (#7286)
## Release notes: product changes

We make authentication layer asynchronous to avoid blocking the async
runtime.

## Motivation

Currently, the authentication layer blocks the runtime worker thread for
as long as it needs to compute the password hash. This happens for every
request, which means the server can only process as many requests at a
time as there are worker threads.

## Implementation

Replace blocking tonic interceptor with an async authentication service.